+++ util/pin-security.ts
@@ -39,6 +39,7 @@
  */
 import { scrypt } from "@noble/hashes/scrypt";
 import { bytesToHex, hexToBytes } from "@noble/hashes/utils";
+import { gcm } from "@noble/ciphers/aes";
 import { getRandomBytes } from "./random";
 import { constantTimeEqual } from "./security-utils";
 import { getValue, saveValue } from "./secure-store";
@@ -64,6 +65,126 @@ const SCRYPT_CONFIG = {
   p: 1, // Parallelization parameter
   dkLen: 32, // Derived key length (256 bits)
 };
+
+/**
+ * Generates a secure key from a PIN using Scrypt (memory-hard function)
+ *
+ * @param pinData - The PIN as Uint8Array
+ * @param salt - The salt for key derivation as Uint8Array
+ * @returns Key suitable for AES encryption
+ */
+function generateKeyFromPin(pinData: Uint8Array, salt: Uint8Array): Uint8Array {
+  // Use Scrypt for memory-hard key derivation with provided salt
+  return scrypt(pinData, salt, SCRYPT_CONFIG);
+}
+
+/**
+ * Encrypts data using a PIN.
+ * Uses AES-GCM for secure encryption with per-record salt.
+ *
+ * @param value - The data to encrypt as Uint8Array
+ * @param pin - The PIN as Uint8Array
+ * @returns The encrypted data as Uint8Array (salt + nonce + ciphertext)
+ */
+function encryptWithPin(value: Uint8Array, pin: Uint8Array): Uint8Array {
+  if (!value || value.length === 0) return new Uint8Array(0);
+
+  try {
+    // Generate a random salt for this encryption (16 bytes)
+    const salt = getRandomBytes(16);
+
+    // Generate a key from the PIN and salt
+    const keyBytes = generateKeyFromPin(pin, salt);
+
+    // Generate a random nonce/IV using our random utility
+    const nonce = getRandomBytes(12); // 12-byte nonce is standard for GCM
+
+    // Use AES-GCM for authenticated encryption
+    const cipher = gcm(keyBytes, nonce);
+    const ciphertext = cipher.encrypt(value);
+
+    // Combine salt + nonce + ciphertext for storage
+    const result = new Uint8Array(
+      salt.length + nonce.length + ciphertext.length,
+    );
+    result.set(salt, 0);
+    result.set(nonce, salt.length);
+    result.set(ciphertext, salt.length + nonce.length);
+
+    return result;
+  } catch (e) {
+    console.error(
+      "Encryption error:",
+      e instanceof Error ? e.message : String(e),
+    );
+    return new Uint8Array(0);
+  }
+}
+
+/**
+ * Decrypts a value that was encrypted with a PIN.
+ * Uses AES-GCM authentication tag for integrity verification.
+ *
+ * @param encryptedValue - The encrypted data as Uint8Array (salt + nonce + ciphertext)
+ * @param pin - The PIN as Uint8Array
+ * @returns An object with the decrypted data and verification status,
+ *          or null if decryption fails
+ */
+function decryptWithPin(
+  encryptedValue: Uint8Array,
+  pin: Uint8Array,
+): { value: Uint8Array; verified: boolean } | null {
+  if (!encryptedValue || encryptedValue.length < 28) {
+    // 16 (salt) + 12 (nonce) minimum
+    return null;
+  }
+
+  try {
+    // Extract salt (first 16 bytes), nonce (next 12 bytes), and ciphertext (rest)
+    const salt = encryptedValue.slice(0, 16);
+    const nonce = encryptedValue.slice(16, 28);
+    const ciphertext = encryptedValue.slice(28);
+
+    // Generate key from PIN and extracted salt
+    const keyBytes = generateKeyFromPin(pin, salt);
+
+    // Create AES-GCM decipher
+    const decipher = gcm(keyBytes, nonce);
+
+    // Decrypt data - AES-GCM will verify integrity automatically
+    let decryptedBytes;
+    try {
+      decryptedBytes = decipher.decrypt(ciphertext);
+    } catch {
+      // AES-GCM authentication failed - wrong PIN or corrupted data
+      return { value: new Uint8Array(0), verified: false };
+    }
+
+    // If we reach here, AES-GCM authentication passed
+    return { value: decryptedBytes, verified: true };
+  } catch (error) {
+    console.error(
+      "Decryption error:",
+      error instanceof Error ? error.message : String(error),
+    );
+    return null;
+  }
+}
+
+/**
+ * Helper functions for string conversion.
+ */
+function stringToUint8Array(str: string): Uint8Array {
+  return new TextEncoder().encode(str);
+}
+
+function uint8ArrayToBase64(array: Uint8Array): string {
+  return btoa(String.fromCharCode(...array));
+}
+
+function base64ToUint8Array(base64: string): Uint8Array {
+  return Uint8Array.from(atob(base64), (c) => c.charCodeAt(0));
+}
 
 /**
  * Hashes a PIN using Scrypt for secure storage and verification.
@@ -254,28 +375,71 @@ export { hashPin };
 export const validatePin = validatePinWithRateLimit;
 export const verifyStoredPin = validatePinWithRateLimit;
 
-// Placeholder exports for functions that were removed/simplified
-// These can be implemented later if needed
+// High-level wrapper functions for data encryption/decryption with PIN
 export async function secureEncryptWithPin(
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  _data: string,
+  data: string,
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  _pin: string,
+  pin: string,
 ): Promise<string | null> {
-  console.warn(
-    "secureEncryptWithPin is not implemented in Phase 1 - use crypto.ts directly",
-  );
-  return null;
+  return processWithPin(pin, async (securePin) => {
+    try {
+      // Convert data to Uint8Array for encryption
+      const dataBytes = stringToUint8Array(data);
+      const pinBytes = stringToUint8Array(securePin);
+
+      // Encrypt using the internal crypto function
+      const encryptedBytes = encryptWithPin(dataBytes, pinBytes);
+
+      if (!encryptedBytes || encryptedBytes.length === 0) {
+        console.warn("Encryption failed - empty result");
+        return null;
+      }
+
+      // Convert to base64 for storage
+      return uint8ArrayToBase64(encryptedBytes);
+    } catch (error) {
+      console.warn("Encryption failed:", error);
+      return null;
+    }
+  });
 }
 
 export async function secureDecryptWithPin(
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  _encryptedData: string,
+  encryptedData: string,
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  _pin: string,
+  pin: string,
 ): Promise<{ value: string; verified: boolean } | null> {
-  console.warn(
-    "secureDecryptWithPin is not implemented in Phase 1 - use crypto.ts directly",
-  );
-  return null;
+  return processWithPin(pin, async (securePin) => {
+    try {
+      // Convert from base64 to Uint8Array
+      const encryptedBytes = base64ToUint8Array(encryptedData);
+      const pinBytes = stringToUint8Array(securePin);
+
+      // Decrypt using the internal crypto function
+      const result = decryptWithPin(encryptedBytes, pinBytes);
+
+      if (!result) {
+        console.warn("Decryption failed - null result");
+        return null;
+      }
+
+      if (!result.verified) {
+        console.warn("Decryption failed - verification failed");
+        return { value: "", verified: false };
+      }
+
+      // Convert back to string
+      const value = new TextDecoder().decode(result.value);
+      return { value, verified: true };
+    } catch (error) {
+      console.warn("Decryption failed - possibly due to incorrect PIN", error);
+      return null;
+    }
+  });
 }
+
+// Export helper functions for compatibility
+export { stringToUint8Array, uint8ArrayToBase64, base64ToUint8Array };
+export { encryptWithPin, decryptWithPin };
